requirements.txt
minio==7.1.14
apache-airflow-providers-docker>=3.0.0,<4.0.0
